Add a new Preload component
