+++ src/Symfony/Bridge/Twig/Extension/AssetExtension.php
@@ -12,7 +12,6 @@
 namespace Symfony\Bridge\Twig\Extension;
 
 use Symfony\Component\Asset\Packages;
-use Symfony\Component\Asset\Preload\PreloadManagerInterface;
 
 /**
  * Twig extension for the Symfony Asset component.
@@ -22,12 +21,10 @@ use Symfony\Component\Asset\Preload\PreloadManagerInterface;
 class AssetExtension extends \Twig_Extension
 {
     private $packages;
-    private $preloadManager;
 
-    public function __construct(Packages $packages, PreloadManagerInterface $preloadManager = null)
+    public function __construct(Packages $packages)
     {
         $this->packages = $packages;
-        $this->preloadManager = $preloadManager;
     }
 
     /**
@@ -38,7 +35,6 @@ class AssetExtension extends \Twig_Extension
         return array(
             new \Twig_SimpleFunction('asset', array($this, 'getAssetUrl')),
             new \Twig_SimpleFunction('asset_version', array($this, 'getAssetVersion')),
-            new \Twig_SimpleFunction('preload', array($this, 'preload')),
         );
     }
 
@@ -69,26 +65,6 @@ class AssetExtension extends \Twig_Extension
     public function getAssetVersion($path, $packageName = null)
     {
         return $this->packages->getVersion($path, $packageName);
-    }
-
-    /**
-     * Preloads an asset.
-     *
-     * @param string $path   A public path
-     * @param string $as     A valid destination according to https://fetch.spec.whatwg.org/#concept-request-destination
-     * @param bool   $nopush If this asset should not be pushed over HTTP/2
-     *
-     * @return string The path of the asset
-     */
-    public function preload($path, $as = '', $nopush = false)
-    {
-        if (null === $this->preloadManager) {
-            throw new \RuntimeException('A preload manager must be configured to use the "preload" function.');
-        }
-
-        $this->preloadManager->addResource($path, $as, $nopush);
-
-        return $path;
     }
 
     /**
