+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.xml
@@ -72,7 +72,10 @@
 
         <service id="twig.extension.assets" class="Symfony\Bridge\Twig\Extension\AssetExtension" public="false">
             <argument type="service" id="assets.packages" />
-            <argument type="service" id="assets.preload_manager" on-invalid="ignore" />
+        </service>
+
+        <service id="twig.extension.preload" class="Symfony\Bridge\Twig\Extension\PreloadExtension" public="false">
+            <argument type="service" id="assets.preload_manager" />
         </service>
 
         <service id="twig.extension.code" class="Symfony\Bridge\Twig\Extension\CodeExtension" public="false">
