+++ src/Symfony/Component/Preload/.gitignore
@@ -1,0 +1,3 @@
+vendor/
+composer.lock
+phpunit.xml
