+++ src/Symfony/Component/Preload/CHANGELOG.md
@@ -1,0 +1,7 @@
+CHANGELOG
+=========
+
+3.3.0
+-----
+
+ * added the component
