+++ src/Symfony/Component/Preload/PreloadManagerInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Asset\Preload;
+namespace Symfony\Component\Preload;
 
 /**
  * Manages resources to preload according to the W3C "Preload" specification.
