+++ src/Symfony/Component/Preload/README.md
@@ -1,0 +1,14 @@
+Preload Component
+===============
+
+The Preload component manages preloading and prefetching of HTTP resources through Link headers
+and HTTP/2 pushes.
+
+Resources
+---------
+
+  * [Documentation](https://symfony.com/doc/current/components/preload/introduction.html)
+  * [Contributing](https://symfony.com/doc/current/contributing/index.html)
+  * [Report issues](https://github.com/symfony/symfony/issues) and
+    [send Pull Requests](https://github.com/symfony/symfony/pulls)
+    in the [main Symfony repository](https://github.com/symfony/symfony)
