+++ src/Symfony/Component/Preload/Tests/PreloadManagerTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Asset\Preload;
+namespace Symfony\Component\Preload;
 
 use PHPUnit\Framework\TestCase;
