+++ src/Symfony/Component/Preload/composer.json
@@ -1,0 +1,39 @@
+{
+    "name": "symfony/preload",
+    "type": "library",
+    "description": "Symfony Preload Component",
+    "keywords": ["HTTP/2", "preload", "prefetch", "push", "performance"],
+    "homepage": "https://symfony.com",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Kévin Dunglas",
+            "email": "dunglas@gmail.com"
+        },
+        {
+            "name": "Symfony Community",
+            "homepage": "https://symfony.com/contributors"
+        }
+    ],
+    "require": {
+        "php": ">=5.5.9"
+    },
+    "suggest": {
+        "symfony/http-kernel": ""
+    },
+    "require-dev": {
+        "symfony/http-kernel": "^2.8 || ^3.0"
+    },
+    "autoload": {
+        "psr-4": { "Symfony\\Component\\Preload\\": "" },
+        "exclude-from-classmap": [
+            "/Tests/"
+        ]
+    },
+    "minimum-stability": "dev",
+    "extra": {
+        "branch-alias": {
+            "dev-master": "3.3-dev"
+        }
+    }
+}
